# feat: implement ultra-performance journal processing system