… ENHANCED_SYSTEM_PERFORMANCE_REPORT.md
… ULTRA_PERFORMANCE_FINAL_REPORT.md
… benchmark_results.json
… enhanced_system/IMPLEMENTATION_REPORT.md
… enhanced_system/TRACEABILITY_GAP_ANALYSIS.md
… enhanced_system/database/04_optimized_functions.sql
… enhanced_system/requirements.txt
… enhanced_system/scripts/performance_comparison.py
… enhanced_system/src/__pycache__/__init__.cpython-313.pyc
… enhanced_system/src/api/routers/journal_router.py
… enhanced_system/src/core/__pycache__/config.cpython-313.pyc
… enhanced_system/src/core/config.py
… enhanced_system/src/core/database.py
… enhanced_system/src/models/database.py
… enhanced_system/src/models/domain.py
… enhanced_system/src/services/__pycache__/enhanced_journal_service.cpython-313.pyc
… enhanced_system/src/services/__pycache__/optimized_journal_service.cpython-313.pyc
… enhanced_system/src/services/audit_service.py
… enhanced_system/src/services/compliance_service.py
… enhanced_system/src/services/enhanced_journal_service.py
… enhanced_system/src/services/journal_service.py
… enhanced_system/src/services/optimized_journal_service.py
… enhanced_system/src/services/voucher_service.py
… run_performance_test.py
… test_enhanced_system.py
… ultra_fast_processor.py
… ultra_fast_results.json
… ultra_performance_processor.py