- Added enhanced PostgreSQL with 21.4% performance improvement via indexes
- Implemented DuckDB + Polars architecture achieving 40% overall speedup
- Created comprehensive audit trail and compliance services
- Identified data I/O bottleneck (92% of processing time)
- Demonstrated 22x processing capability when data in memory
- Added FastAPI service layer with async processing
- Implemented fraud detection using Benford's Law
- Created detailed performance analysis and recommendations

Performance Results:
- Original: 3,023 ms
- With Indexes: 2,376 ms (21% faster)
- DuckDB+Polars: 1,823 ms (40% faster)
- Processing only: 134 ms (22x faster potential)

Technologies: PostgreSQL, DuckDB, Polars, FastAPI, SQLAlchemy

🤖 Generated with Claude Code